+++ openspec/changes/civ7-support-direct-control-modularization/workstream/workstream-record.md
@@ -627,6 +627,18 @@ All future agent waves must be framed before delegation:
   projection, Effect/oRPC procedure-core work, or Task 2.9.4 matrix-row
   acceptance. This is local package/type relocation proof only, not runtime
   proof, AI/hotseat product-path support, or procedure-core readiness.
+- Direct-control summary read types slice: completed as a narrow public type
+  ownership relocation. It moves only player, unit, and city summary
+  input/result types into `src/play/summaries.ts` while keeping facade type
+  re-exports in `index.ts` and replacing the summary module's type-only facade
+  imports with concrete owner imports. This preserves public type contracts and
+  does not change summary validation, component-id pass-through behavior,
+  source strings, runtime behavior, relationship-label policy, visibility or
+  GameInfo types, procedure schemas, telemetry, AI ingestion, hotseat runtime
+  proof, CLI semantic projection, Effect/oRPC procedure-core work, or Task
+  2.9.4 matrix-row acceptance. This is local package/type relocation proof
+  only, not runtime proof, AI/hotseat product-path support, or procedure-core
+  readiness.
 - Direct-control turn-completion slice: completed as a narrow turn-completion
   source/wrapper relocation. It moves only `getCiv7TurnCompletionStatus`,
   `sendCiv7TurnComplete`, `sendCiv7TurnUnready`, the status command builder,
